+++ quilt/src/main/java/com/faboslav/friendsandfoes/quilt/EarlyRiserQuilt.java
@@ -1,6 +1,7 @@
 package com.faboslav.friendsandfoes.quilt;
 
 import com.chocohead.mm.api.ClassTinkerers;
+import com.chocohead.mm.api.EnumAdder;
 import com.faboslav.friendsandfoes.FriendsAndFoes;
 import com.faboslav.friendsandfoes.init.ModEntity;
 import com.faboslav.friendsandfoes.util.ExpandedEnumValues;
@@ -16,11 +17,18 @@ public final class EarlyRiserQuilt implements Runnable
 		String Raid = remapper.mapClassName("intermediary", "net.minecraft.class_3765$class_3766");
 		String EntityType = 'L' + remapper.mapClassName("intermediary", "net.minecraft.class_1299") + ';';
 
-		ClassTinkerers
-			.enumBuilder(Raid, EntityType, int[].class)
-			.addEnum(ExpandedEnumValues.ILLUSIONER, () -> new Object[]{net.minecraft.entity.EntityType.ILLUSIONER, new int[]{0, 0, 0, 0, 1, 0, 1, 1}})
-			.addEnum(ExpandedEnumValues.ICEOLOGER, () -> new Object[]{ModEntity.ICEOLOGER.get(), new int[]{0, 0, 0, 0, 1, 1, 0, 1}})
-			.build();
+		EnumAdder builder = ClassTinkerers
+			.enumBuilder(Raid, EntityType, int[].class);
+
+		if(FriendsAndFoes.getConfig().enableIllusionerInRaids) {
+			builder.addEnum(ExpandedEnumValues.ILLUSIONER, () -> new Object[]{net.minecraft.entity.EntityType.ILLUSIONER, new int[]{0, 0, 0, 0, 1, 0, 1, 1}});
+		}
+
+		if(FriendsAndFoes.getConfig().enableIceologerInRaids) {
+			builder.addEnum(ExpandedEnumValues.ICEOLOGER, () -> new Object[]{ModEntity.ICEOLOGER.get(), new int[]{0, 0, 0, 0, 1, 1, 0, 1}});
+		}
+
+		builder.build();
 
 		String SpawnGroup = remapper.mapClassName("intermediary", "net.minecraft.class_1311");